Refactored buffer streams
